Predict an object of class modeler

Generate model predictions from an object of class modeler. This function allows for flexible prediction types, including point predictions, area under the curve (AUC), and first or second order derivatives.

Usage

# S3 method for class 'modeler'
predict(
  object,
  x = NULL,
  id = NULL,
  type = c("point", "auc", "fd", "sd"),
  se_interval = c("confidence", "prediction"),
  n_points = 1000,
  metadata = FALSE,
  ...
)

Arguments

object

An object of class modeler, typically the result of calling the modeler() function.

x

A numeric value or vector specifying the points at which predictions are made. For type = "auc", x must be a vector of length 2 that specifies the interval over which to calculate the AUC.

id

Optional unique identifier to filter predictions by a specific group. Default is NULL.

type

A character string specifying the type of prediction. Default is "point".

"point"

Predicts the value of y for the given x.

"auc"

Calculates the area under the curve (AUC) for the fitted model over the interval specified by x.

"fd"

Returns the first derivative (rate of change) of the model at the given x value(s).

"sd"

Returns the second derivative of the model at the given x value(s).

se_interval

A character string specifying the type of interval for standard error calculation. Options are "confidence" (default) or "prediction".

n_points

An integer specifying the number of points used to approximate the area under the curve (AUC) when type = "auc". Default is 1000.

metadata

Logical. If TRUE, metadata is included with the predictions. Default is FALSE.

...

Additional parameters for future functionality.

Value

A data.frame containing the predicted values, their associated standard errors, and optionally the metadata.

Author

Johan Aparicio [aut]

Examples

library(flexFitR)
data(dt_potato)
mod_1 <- dt_potato |>
  modeler(
    x = DAP,
    y = Canopy,
    grp = Plot,
    fn = "fn_piwise",
    parameters = c(t1 = 45, t2 = 80, k = 0.9),
    subset = c(15, 2, 45),
    options = list(add_zero = TRUE, max_as_last = TRUE)
  )
print(mod_1)
#> 
#> Call:
#> Canopy ~ fn_piwise(DAP, t1, t2, k) 
#> 
#> Sum of Squares Error:
#>     Min.  1st Qu.   Median     Mean  3rd Qu.     Max. 
#> 0.002601 0.504829 1.007057 2.674719 4.010777 7.014498 
#> 
#> Optimization Results `head()`:
#>  uid   t1   t2     k    sse
#>    2 35.2 60.7 100.0 7.0145
#>   15 38.5 69.4  99.8 1.0071
#>   45 38.3 64.5 100.0 0.0026
#> 
#> Metrics:
#>  Groups      Timing Convergence Iterations
#>       3 1.3998 secs        100%   414 (id)
#> 
# Point Prediction
predict(mod_1, x = 45, type = "point", id = 2)
#> # A tibble: 1 × 4
#>     uid x_new predicted.value std.error
#>   <dbl> <dbl>           <dbl>     <dbl>
#> 1     2    45            38.4     0.622
# AUC Prediction
predict(mod_1, x = c(0, 108), type = "auc", id = 2)
#> # A tibble: 1 × 5
#>     uid x_min x_max predicted.value std.error
#>   <dbl> <dbl> <dbl>           <dbl>     <dbl>
#> 1     2     0   108           6004.      30.8
# First Derivative
predict(mod_1, x = 45, type = "fd", id = 2)
#> # A tibble: 1 × 4
#>     uid x_new predicted.value std.error
#>   <dbl> <dbl>           <dbl>     <dbl>
#> 1     2    45            3.91    0.0741
# Second Derivative
predict(mod_1, x = 45, type = "sd", id = 2)
#> # A tibble: 1 × 4
#>     uid x_new predicted.value std.error
#>   <dbl> <dbl>           <dbl>     <dbl>
#> 1     2    45   -0.0000000298 0.0000274